Added tests for MFC class.
Added some test conditions for the MFC class. Added a getter for the
time value for the MFC's ODE. Fixed the fingerprint of the
first_order_delay function. In running tests, I also plotted the unit
step response of an MFC and found that it is not behaving correctly.
Need to look into the Pade approximation of a first-order ODE with time
delay some more, or another way to simulate this.