Began adding tests for Controller class.
Began adding tests for the Controller class and updating the methods
therein. Need to test the update method and control law, but I suppose I
also need a better way of assigning a control law for each MFC and a
better control law than the placeholder that is there currently. Going
to mess with branching so that these WIP commits are not in the master
branch.