Fixed the first-order time delay ODE and added output function.
Added an output function to the MFC object that can be thought of as a
measurement  function for the state update equation defined by the MFC's
ODE. Also fixed the second-order Pade approximation of a first-order
system with time delay and verified that it behaves as expected. These
were test driven, so the tests reflect these changes. Made a bit more
progress on implementing tests for the Controller class, but mostly
fixed the issues with the MFC's ODE in the main lab_hardware.py file.