Added separate methods to get MFC state and output.
Modified/added getters for MFC so that the internal state of the ODE can
be returned with get_state(), and the output of the ODE through the
MFC's output function can be returned with get_output(). Refactored the
code to use get_output() for the existing tests. I think I will end up
implementing a state space controller on the internal state of the ODE
inside of the Controller class.