Added handling of single-MFC Controller instance.
Added some checks and handling to the constructor and getters for the
Controller class so that it can run successfully if initialized with a
single MFC (no longer assumes that you are passing a list of MFCs).
Could add plenty more checks and exceptions for to ensure that the
controller is initailized with objects of the correct type. Maybe later,
once I get the core functionality working.