Refactored linear models to use matrices, added KalmanFilter class.
Refactored first_order_dely and first_order_output so that they really
just generalize a state space equation. Updated tests to define these
equations with their A, B, and C matrices. All of this was to make it
easier to define a KalmanFilter class using the same state-space model.
Started making tests for the KalmanFilter class and also implementing
the Kalman filter algorithm for a linear state space model (nothin
nonlinear yet).