Fixed the Kalman filter! With tests.
I think my code coverage with my tests is pretty poor, but they have
helped me debug and refactor successfully, so I call them good enough
for now. Got the KF working properly. Needed to change the data types
that some functions return to do it, and mess around a bit with
reshaping arrays. It seems scipy.integrate.ode likes to change its state
to a single-row array even if it is stored as a column-array, which
causes some problems with dot products. All good now! The KF's
performance isn't the greatest though, still may need to tune it or go
with something more sophistocated.