Moved initialization code to main.py, and added get_state_matrices() method to sim_functions.py.
Moved code for initializing the different components of the Combustor
class into main.py, so the constructor for this class now basically just
assigns these components, and the specifics for our system are set in
main. Also added a helper function to sim_functions that takes
first-order system gain, time constant, and time delay, and returns the
A, B, and C state space matrices for the system using a second-order
Pade approximation of the time delay. Updated tests to use this
function, and added tests to verify that the function works as expected.