Added actual dynamic pressure sensor locations.
Added actual locations of the dynamic pressure sensors on the lab
apparatus, measured in mm from the base of the combustion chamber (not
the base of the flame!). Also updated the initialization portion of
main.py to add a single DynamicSensor object to sensor_list for each of
these sensor locations.