Added ODE model to dynamic sensor.
Added code to initilaize an ODE model passed to the dynamic sensor and
update its dynamics at the specified sensor rate for each call to the
dynamic sensor. Now I just need to figure out a model that makes sense
for my simulation and test it. Again, did the code before the tests. Bad
programmer!