Added actual MFC static gains (SLPM/V) and time delays (sec) to main.py.
Also, it looks like the middle fuel MFC is as first-order as it gets.
The other two do some funky stuff that I might be able to capture
decently with either a first- or second-order model, but I think the
valve is sticking on them. Disappointing for brand new MFCs...